examples/mercedes_benz_automobile/backend/pyproject.toml
[project]
name = "chatkit-mercedes-benz-automobile"
version = "0.1.0"
requires-python = ">=3.12"
description = "Mercedes-Benz Automobile Voice Assistant - ChatKit Example"
readme = "README.md"
dependencies = [
    "fastapi>=0.114.1,<0.116",
    "uvicorn>=0.30,<1",
    "openai>=1.40,<2",
    "openai-chatkit>=1.0.2,<2",
    "agents>=0.5.0,<1",
    "httpx>=0.28,<1",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.5.5",
    "mypy>=1.11.1",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []
